Change to float

+++ mgraph.c
@@ -5,27 +5,27 @@
 struct m_graph {
     int dir; // direcionado ou nao
     int n;  // tamanho (num de vertices)
-    int** matrix;
+    float** matrix;
 };
 
 M_GRAPH* m_graph_create(int n, int dir) {
     M_GRAPH* g = (M_GRAPH*) malloc(sizeof(M_GRAPH));
     int i, j;
-    
+
     g->n = n;
     g->dir = dir;
-    g->matrix = (int **) malloc(sizeof(int *) * n);
+    g->matrix = (float **) malloc(sizeof(float *) * n);
     for (i = 0; i < n; i++)
-        g->matrix[i] = (int *) malloc(sizeof(int) * n);
-    
+        g->matrix[i] = (float *) malloc(sizeof(float) * n);
+
     for (i = 0; i < n; i++)
         for (j = 0; j < n; j++)
             g->matrix[i][j] = -1; // matrix vazia
-    
+
     return g;
 }
 
-void m_graph_insert(M_GRAPH* g, int row, int col, int w) {
+void m_graph_insert(M_GRAPH* g, int row, int col, float w) {
     g->matrix[row][col] = w;
     if (!g->dir) g->matrix[col][row] = w;
 }
@@ -37,7 +37,7 @@ int m_graph_checkedge(M_GRAPH* g, int row, int col) {
 
 void m_graph_adj(M_GRAPH* g, int v) {
     int i;
-    
+
     for (i = 0; i < g->n; i++) {
         if (g->matrix[v][i] >= 0)
             printf("%d ", i); // vertices adjacentes
@@ -59,7 +59,7 @@ void m_graph_print(M_GRAPH* g) {
     for (i = 0; i < g->n; i++) {
         for (j = 0; j < g->n; j++) {
             if (g->matrix[i][j] < 0) printf(". ");
-            else printf("%d ", g->matrix[i][j]);
+            else printf("%f ", g->matrix[i][j]);
         }
         printf("\n");
     }
@@ -68,29 +68,29 @@ void m_graph_print(M_GRAPH* g) {
 M_GRAPH* m_graph_transpose(M_GRAPH* g) {
     M_GRAPH* t = m_graph_create(g->n,g->dir);
     int i, j;
-    
+
     for (i = 0; i < t->n; i++)
         for (j = 0; j < t->n; j++)
             t->matrix[j][i] = g->matrix[i][j]; // transposicao
-    
+
     return t;
 }
 
 void m_graph_printtranspose(M_GRAPH* g) {
     M_GRAPH* t = m_graph_create(g->n,g->dir);
     int i, j;
-    
+
     for (i = 0; i < t->n; i++)
         for (j = 0; j < t->n; j++)
             t->matrix[j][i] = g->matrix[i][j]; // transposicao
-    
+
     m_graph_print(t);
     m_graph_free(t);
 }
 
 void m_graph_printloweredge(M_GRAPH* g) {
     int i, j, row, col, lower = MAX;
-    
+
     for (i = 0; i < g->n; i++) {
         for (j = 0; j < g->n; j++) {
             if (g->matrix[i][j] < lower && g->matrix[i][j] >= 0) {
@@ -99,7 +99,7 @@ void m_graph_printloweredge(M_GRAPH* g) {
             }
         }
     }
-    
+
     if (!g->dir) { // printa a aresta com menor peso corretamente
         if (row <= col) {
             printf("%d %d\n", row, col);
